requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.7.4.post0
APScheduler==3.9.1
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.2
colorama==0.4.5
commonmark==0.9.1
cryptography==36.0.0
cupy==11.0.0
cycler==0.11.0
Cython==0.29.30
fastrlock==0.8
fonttools==4.32.0
gevent==21.12.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.42.0
html5lib==1.1
idna==3.3
imageio==2.17.0
importlib-metadata==4.11.3
kiwisolver==1.4.2
lxml==4.8.0
Markdown==3.3.6
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcls==0.23.2
mmcv-full==1.5.3
model-index==0.1.11
multidict==5.2.0
networkx==2.8
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.64
openmim==0.2.1
ordered-set==4.1.0
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
pip==21.2.4
prettytable==3.3.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0.4
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
setuptools==61.2.0
Shapely==1.8.2
siren-torch==1.1
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.2
termcolor==1.1.0
tifffile==2022.4.8
torch==1.11.0
torchsummaryX==1.3.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.8
watch==0.2.7
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.3.2
Werkzeug==2.1.1
wheel==0.37.1
yapf==0.32.0
yarl==1.6.3
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
